[UPDATE] 新增手机注册 等一些常用方法

+++ mkshop-manager/mkshop-manager-interface/src/main/java/com/nexus/manager/service/UserService.java
@@ -1,6 +1,7 @@
 package com.nexus.manager.service;
 
 
+import com.nexus.common.model.ServerResponse;
 import com.nexus.manager.pojo.TbUser;
 
 import java.util.ArrayList;
@@ -10,7 +11,7 @@ public interface UserService {
 
     public Integer insert_user(TbUser tbUser);
 
-    public TbUser select_user(Integer id);
+    public TbUser select_user(long id);
 
     public TbUser selectUserByName(String username);
