+++ mkshop-manager/mkshop-manager-service/src/main/java/com/nexus/manager/service/impl/UserServiceImpl.java
@@ -1,13 +1,25 @@
 package com.nexus.manager.service.impl;
 
 
+import com.nexus.common.cache.TokenCache;
+import com.nexus.common.model.ResponseCode;
+import com.nexus.common.model.ServerResponse;
+import com.nexus.common.utils.MessageUtil;
+import com.nexus.common.utils.RandomUtil;
 import com.nexus.manager.mapper.TbUserMapper;
 import com.nexus.manager.pojo.TbUser;
 import com.nexus.manager.service.UserService;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.DigestUtils;
+
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
 
 
 @Service("userService")
@@ -24,7 +36,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public TbUser select_user(Integer id) {
+    public TbUser select_user(long id) {
         return tbUserMapper.selectByPrimaryKey(id);
     }
 
@@ -37,6 +49,7 @@ public class UserServiceImpl implements UserService {
     public Integer updateUser(TbUser tbUser) {
         return tbUserMapper.updateByPrimaryKeySelective(tbUser);
     }
+
 
 
 }
